Rename test classes
