+++ library/runtime-core/src/nonJsTest/kotlin/io/matthewnelson/kmp/tor/runtime/core/util/JobUtilNonJsUnitTest.kt
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package io.matthewnelson.kmp.tor.runtime.core
+package io.matthewnelson.kmp.tor.runtime.core.util
 
 import io.matthewnelson.kmp.file.IOException
 import io.matthewnelson.kmp.tor.core.api.annotation.InternalKmpTorApi
+import io.matthewnelson.kmp.tor.runtime.core.EnqueuedJob
 import io.matthewnelson.kmp.tor.runtime.core.EnqueuedJobUnitTest.TestJob
-import io.matthewnelson.kmp.tor.runtime.core.util.awaitSync
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.IO
 import kotlinx.coroutines.delay
@@ -30,7 +30,7 @@ import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.TimeSource
 
 @OptIn(InternalKmpTorApi::class)
-class EnqueuedJobNonJsUnitTest {
+class JobUtilNonJsUnitTest {
 
     @Test
     fun givenAwaitSync_whenCancellationLambdaThrows_thenIsWrappedInCancellationException() {
